mapreduce-reference/requirements.txt
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.1
pytest==7.4.3
pytest-timeout==2.2.0
matplotlib==3.8.2
numpy==1.26.2
